refactor: site structure change
-サイト構造変更。headerを付ける付けないでのディレクトリ構造を変更
