+++ app/challenge/[competitionId]/[umpireId]/[playerId]/@modal/modal.tsx
