+++ app/challenge/[competitionId]/[umpireId]/[playerId]/@modal/page.tsx
