+++ app/challenge/[competitionId]/[umpireId]/[playerId]/loading.tsx
