+++ app/challenge/[competitionId]/[umpireId]/view.tsx
