+++ app/challenge/challengeModal.tsx
