+++ app/challenge/page.tsx
@@ -1,6 +1,6 @@
 import { getPlayerList, getCourseList } from "@/app/components/common/utils"
 import type { SelectCourse, SelectPlayer } from "@/app/lib/db/schema"
-import { View } from "@/app/(nolinkheader)/challenge/view"
+import { View } from "@/app/challenge/view"
 
 export default async function Challenge() {
   const courseDataList: { selectCourses: SelectCourse[] } = await getCourseList()
