+++ app/challenge/view.tsx
@@ -3,9 +3,9 @@ import Link from "next/link"
 import { useState } from "react"
 import useSound from "use-sound"
 import type { SelectCourse, SelectPlayer } from "@/app/lib/db/schema"
-import ChallengeList from "@/app/(nolinkheader)/challenge/challengeList"
+import ChallengeList from "@/app/challenge/challengeList"
 import { CommonRadioList } from "@/app/components/common/commonList"
-import Challenge from "@/app/(nolinkheader)/challenge/challenge"
+import Challenge from "@/app/challenge/challenge"
 import CircleButton from "@/app/components/parts/circleButton"
 import { SensorCourse } from "@/app/components/challenge/sensorCourse"
 import StartSound from "@/app/lib/sound/01_start.mp3"
