+++ app/config/assignList/page.tsx
@@ -1,6 +1,6 @@
 import { SelectAssignList } from "@/app/lib/db/schema"
 import { getAssignList } from "@/app/lib/db/queries/queries"
-import { View } from "@/app/(linkheader)/config/assignList/view"
+import { View } from "@/app/config/assignList/view"
 
 export const revalidate = 0
