+++ app/config/assignList/view.tsx
