+++ app/course/@modal/assign/[[...courseId]]/page.tsx
