+++ app/course/@modal/default.tsx
