+++ app/course/@modal/delete/[[...courseId]]/page.tsx
