+++ app/course/edit/layout.tsx
