+++ app/course/layout.tsx
