+++ app/course/page.tsx
