+++ app/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next"
 import { Inter } from "next/font/google"
+import { Header } from "@/app/components/header"
 import "./globals.css"
 
 const inter = Inter({ subsets: ["latin"] })
@@ -20,6 +21,7 @@ export default function RootLayout({
     <html lang="ja" className={inter.className}>
       <body className="font-zenKakuGothicNew">
         <main className="mx-auto text-xs sm:px-12 lg:text-base w-screen h-screen">
+          <Header />
           {children}
           {auth}
         </main>
