+++ app/player/@modal/assign/[[...playerId]]/page.tsx
