+++ app/player/@modal/delete/[[...playerId]]/page.tsx
