+++ app/player/layout.tsx
