+++ app/summary/[...ids]/tCourseTable.tsx
