+++ app/summary/[competitionId]/summaryTable.tsx
