+++ app/umpire/@modal/delete/[[...umpireId]]/page.tsx
